fix: Eval{,Full} runs in query monad
Prior to this, evaluation happened in `MonadEditApp`, and thus could
change the fresh ID state embedded in the `App`. This can cause "lost
ID" errors when interleaving actions, evaluations and undo/redos.

Note that it is suspicious that `evalFull'` is still using
`liftEditAppM` and generates a new ID via `gvar` -- this indeed leads to
the same sort of issues, and will be dealt with in a subsequent
commit. (The fact that this code will be changing very soon is why we
have not gone to the effort of using `readerToState` here.)

We have added a test showing that we have fixed this "lost ID" issue at
the "richly-typed api" level. (Note that `evalFull'` is not part of this
layer.)

Signed-off-by: Ben Price <ben@hackworthltd.com>