refactor: move the Primer API into its own package
We would like to run evaluations in the browser, using GHC 9.6's
support for Wasm and JavaScript targets. We won't need the API wrapper
for that, nor type-checking, so this is the first step towards a more
fine-grained set of packages in order to reduce the amount of code we
need to compile for a Wasm/JavaScript target.

This is a pure refactor. No tests or functionality have been removed,
nor any new functionality or tests added. We've simply shuffled things
around a bit, and moved some additional code into `primer-testlib` so
that it can be reused in `primer-api`'s tests.

Signed-off-by: Drew Hess <src@drewhess.com>
Signed-off-by: Ben Price <ben@hackworthltd.com>