feat: eval elides lets as soon as possible
Before doing this, we would sometimes build up large redundant
substitutions, and have to push these down to the leaves before eliding
them. This leads to larger terms and extra work.

This enables us to reenable unit_8, as it gives a major performance
increase. It also noticably speeds up unit_9.

Signed-off-by: Ben Price <ben@hackworthltd.com>