refactor! make eval-under-binders an option with no alternatives
This is a stepping-stone to making an option to not evaluate under
binders for EvalFull.

BREAKING CHANGE: This adds a new field to `EvalFullReq`, which is
exposed in the APIs. Since it is an extra optional argument in the
OpenAPI, old clients of the OpenAPI will still continue to work (this
is not true of the richly-typed API).

Signed-off-by: Ben Price <ben@hackworthltd.com>