feat: Add IDs to kind nodes
We take advantage of this to output proper universally-unique IDs in `API.viewTreeKind`.

We still use `Kind' ()` in the definition of `TForall`, since adding the ability to edit such kinds is a complicated change which is out of scope for now. We're initially focused on kinds in typedefs.

NB. Even though we are considering getting rid of IDs entirely, we'd almost certainly still want some sort of metadata on nodes, even just for caching, and thus we'd benefit from the significant majority of the changes in this commit, since it's mostly concerned with giving the option to attach metadata to kind nodes. In other words, the exact nature of the metadata on kinds is irrelevant to most of the lines which are changed here.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>