feat: evaluation optionally treats adjacent `let`s as a group
We often build terms with many adjacent `let`s:
  let x=... in let y=... in let z=... in f s
We now can push all three of these `let` into the application in one step,
rather than having to push them one-at-a-time. Note that the evaluation
order ensures that (in the old mode) if we take the step of pushing `z`,
then the next redex to reduce is pushing `y`, and the next is pushing
`x`. Thus this option will not affect evaluation order.

This change happens to reduce the required number of steps for
Prelude.Integer.tasty_min_prop and Prelude.Integer.tasty_max_prop enough
to undo the temporary bump.

Signed-off-by: Ben Price <ben@hackworthltd.com>