feat: evaluation optionally elides lets as soon as possible
Before doing this, we would sometimes build up large redundant
substitutions, and have to push these down to the leaves before eliding
them. This leads to larger terms and extra work.

This enables us to reenable unit_8 and unit_9, and primer-benchmark's
"mapEven 10" as it gives a major performance increase.

Whilst this reduces the number of steps taken for most examples,
note that the step limit had to be increased on unit_prim_partial_map,
and unit_type_preservation_BETA_regression also took more steps. These
are both because of the same phenomenon caused by the interaction with
treating adjacent lets as a group, causing more steps to be taken on
smaller trees due to early partial elision. Note that the optimisation
makes no difference on terms like
  let x1=_ in ... let xn=_ in xi
as `xi` (an occurrence of one of the let-bound variables) is a leaf. The
workload in these two tests has many such subterms which get created when
pushing lets down. Aggressive elision will remove some of the lets early:
for example if there was a term such as
  let x1=_ in ... let xn=_ in C xi xj
then aggressively eliding would remove all but two of the lets (assuming
they are independent) as a first step, leaving
  let xi=_ in let xj=_ in C xi xj
but then follow all the steps that the non-aggressive eliding option
would take (push all lets under the constructor, elide around the leaf
`xi`, inline `xi`, elide around `xj`, inline `xj`). Note that this means
that aggressively eliding lets takes an extra step for this pattern.

Signed-off-by: Ben Price <ben@hackworthltd.com>