update testsuite for aggressive elision. some justifications:
bump limit on unit_prim_partial_map because...

it turns out that on this workload the aggressive elision optimisation
actually takes more step. However, the trees are somewhat smaller.
this is because the optimisation makes no difference on terms like
  let x1=_ in ... let xn=_ in xi
as `xi` is a leaf. This workload has many such subterms which get created when pushing lets down. Aggressive elision will remove some of the lets early, when they are (for example)
  let x1=_ in ... let xn=_ in C xi xj
but this takes an extra step

This phenomenon is also seen in unit_type_preservation_BETA_regression