refactor: add parameter for kind meta to Expr' and Type'
This is usually instantiated to `()`, matching the previous hardcoded
choice. There are intended to be no functional changes in this commit,
it is simply setup for adding IDs to kinds in foralls (and thus to enable
actions in those positions).

(Note that there are a couple of TODOs added in comments which will be
addressed in a subsequent commit.)

Signed-off-by: Ben Price <ben@hackworthltd.com>