refactor: remove some typedef traversals
This requires reworking `forgetTypeDefMetadata` and `generateTypeDefIDs`,
where we can use generic optics instead of the current traversals.

We do this in preparation for adding metadata to kinds in types. This
will include kinds within fields of constructors. When that happens, the
same type parameter will control the metadata in two separate fields of
ASTTypeDef, meaning the old type-changing lens-style definitions would
become very awkward.

Signed-off-by: Ben Price <ben@hackworthltd.com>