feat: (unused) machinery to express Kind selections
We add a `Void` field to such selections so that it is clear we don't
yet create them, and it is easy to stub out some codepaths that do not
yet make sense (mostly due to kinds not having IDs yet).

Note that `findTypeOrKind` (and friends) only work for the kind parameter
being `()`, and consequently do not actually focus on kinds (they have
no IDs).

Signed-off-by: Ben Price <ben@hackworthltd.com>