fix: outlaw duplicate binders in case branch
Accepting terms like `case e of C x x -> t` was never intended, and the
semantics are unclear. (This term is a linear pattern with two pattern
variables that just happen to be named the same, so one shadows the other
(it happens that the implementation is that the right-most "wins").
However, this term could easily be expected to be a non-linear match,
i.e. something like `case e of C x y | x == y -> t`.)

Signed-off-by: Ben Price <ben@hackworthltd.com>