feat: eval optionally avoids more shadowing
We already ensure we do not capture variables, so we can easily extend the
logic to avoid shadowing also.  Note that (if evaluating under a binder)
we will still create shadowing when inlining globally-defined things,
namely global definitions and case-of-known-constructor (which inlines
parts of the type definition).  However, when not going under binders,
we should now never introduce shadowing.

Signed-off-by: Ben Price <ben@hackworthltd.com>