test: Primitive functions can exhibit laziness
We could add other tests that would tightly capture the sorts of laziness that our current implementation is capable of. Such as that with `bar x True = 0; bar x False = ...`, evaluating `bar (let x = x in x) True ~> 0` will work, but `bar (let x = x in x) (not False)` will diverge as we reduce the first child before the second. But the tests added here capture the key behaviour that we really care about. And, given that our implementation does not allow for defining primitives locally in the test suite, we're a little reluctant to pollute `PrimDef` with too many constructors which are designed only for testing.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>