fix: CreateDef/newEmptyProgImporting fills type/kind caches
This fixes a problem that newly-created definitions (including the default
"main" definition in a new program) do not offer the correct actions. In
particular, they don't offer a "match" action until something forces
typechecking and their typecaches get correctly filled. (This "something"
is often changing the type of some potentially unrelated definition,
which currently causes a full typechecking run of all definitions.)

Another option would be to run TC on the new definition, but that is
slightly awkward: its type says it may throw an error, but we know in
this situation it will not.

Signed-off-by: Ben Price <ben@hackworthltd.com>