feat: better errors if Eval.step called on a kind/binding
This is a somewhat unusual error case, as presumably the caller would
have some reason to believe the node it is pointing at is actually
a redex. However, we try to give sensible errors even in this case
(in particular, this will be useful for a test in a subsequent
commit). Previously we would return `NodeNotFound` for the `ID` of the
`Type` in `? : ∀a:Type. a`, which is rather misleading!

Signed-off-by: Ben Price <ben@hackworthltd.com>