feat!: expose "expected" and "got" mismatch metadata in api
This simply gives more information in the api `selection/`
(i.e. `getTypeOrKind` operation). We give both the type assigned to the
hole itself as well as the type assigned to the contents. This doesn't
actually expose more information, since one could always call the old api
on both the hole and its contained node, but it is much more convenient
for a consumer to have all the relevant information to understand a
mismatch in one place.

BREAKING CHANGE: this changes the return type of an endpoint in the openapi.

Signed-off-by: Ben Price <ben@hackworthltd.com>