chore: switch from `uuid` to `uuid-types`
`uuid` drags in `network-info`, which doesn't build on GHC's Wasm
target. The only drawback to `uuid-types` for our purposes is that it
doesn't include a cryptographically secure generator for UUIDv4, which
we've been using up until now (specifically, the `nextRandom` effect).

Therefore, we replace `nextRandom` with our own pseudo-random
implementation. Note that our implementation is *not*
cryptographically secure, as `entropy` does not build on the Wasm
target, either. This is fine, as we don't rely on the cryptographic
properties of v4 UUIDs.

Signed-off-by: Drew Hess <src@drewhess.com>