chore: disable ghcjs impl checks in Cabal files
Previously, we checked for GHCJS and disabled tests accordingly. I
think we're unlikely to use GHCJS at this point, so we remove these
checks.

I've also updated a related comment on why we add orphan instances for
some OpenAPI 3 instances, because I believe they're necessary for Wasm
support, in addition to GHCJS.

Signed-off-by: Drew Hess <src@drewhess.com>