feat: Output separate nodes for variable bindings
Note that we currently create fake IDs for these on the fly, as we already did for various elements of patterns. As a result, bindings are in effect not selectable, and thus cannot have actions assigned to them. We will need to give bindings IDs in the core of the library in order to solve this, most likely by making more use of our `Bind` type.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>