fix: make Wasm upstream overrides dependent on `arch(wasm32)`
This fixes an issue where Nix caching was significantly degraded in
the default Nix shell, due to haskell.nix not being able to build
`source-repository-package`s using Nix in Nix shells. We'd prefer not
to do this, because it means that our Wasm dependencies are ever so
slightly different than our native target dependencies, but at the
moment the differences should be small enough not to matter.

Signed-off-by: Drew Hess <src@drewhess.com>