refactor: generalise TestM to a transformer, now have MonadIO WT
This is needed to be able to run `IO` actions in `PropertyWT`, which
will itself be needed when we implement a more efficient evaluator
(due to using `System.Timeout` to avoid infinite loops).

Another choice would be to generalise `WT` to be a transformer
also. However, we only actually use the (hypothetical) `WT IO` and
not any other `WT m`. Since testing showed that the extra polymorphism
caused GHC to emit much less (memory) efficient code, we will use the
monomorphic version. (I expect that sufficient use of specialization
pragmas would fix the memory leaks for a transformer version of `WT`,
but I have not verified this.)

Signed-off-by: Ben Price <ben@hackworthltd.com>