chore: always build `wasm32` targets with `-O2`
As Wasm is an execute-only target, it makes sense to always build the
Primer Wasm target (and its dependencies) with `-O2`. We already do
this for Nix CI builds for non-Wasm targets, but we don't currently
build Wasm targets with Nix, so we use `cabal.project` to specify this
for Wasm.

(If it's desirable to disable this for local development work, it can
be overridden in `cabal.project.local`, I believe.)

Fixes https://github.com/hackworthltd/primer/issues/1196

Signed-off-by: Drew Hess <src@drewhess.com>