fix(primer-service): update OpenAPI golden test output
The previous commit bumped the `openapi3` package to version `3.2.4`,
which contains this pull request:

https://github.com/biocad/openapi3/pull/88

The title of that PR is simply, "Give title to sub-schemas of oneOf
for sum types", and that is what has happened here to our golden test
output. Unhelpfully, there's no justification for why this change was
made, but it looks harmless enough. (This will presumably require a
corresponding schema change on the client side, though.)

We also bump the PVP for `openapi3` to reflect the new minimum version
requirement.

Signed-off-by: Drew Hess <src@drewhess.com>