fix(wasm32): discover wasm32 tests upon entry to Nix shell
Now that we force-enable tests in `cabal.project`, we must ensure that
the wasm32 target's Tasty tests are baked out prior to running Cabal's
build step. As we use the `wasm` Nix develop shell for wasm32
builds (both interactively and in CI), so we can do this via a shell
hook.

Signed-off-by: Drew Hess <src@drewhess.com>