chore(benchmarks): back to Nix-style benchmarking
Due to some changes and greater efficiencies in our CI system, we can
now rely again on Nix's built-in benchmarking support; that is, use
the `requiredSystemFeatures = ["benchmark"]` Nix property on
benchmarking derivations to guarantee that benchmarks will only run on
benchmark-capable hosts.

For posterity, the more flexible Buildkite queue-based benchmark
system *did* work, but it's no longer necessary, and doing things the
Nix way has a few advantages; namely that the configuration is
simpler, and that Nix will run benchmark jobs as soon as they're
ready. The Buildkite approach we used waited for all the other builds
to finish before kicking off the benchmarks, even if the benchmark's
dependencies had already been built.

Signed-off-by: Drew Hess <src@drewhess.com>