refactor: do some DRY for full eval tests
We refactor many of the interpreter and step evaluator full evaluation
tests by moving the programs we want to evaluate into
`primer-testlib`. Note that not all tests are amenable to this
refactoring: some are too implementation-specific, and some are so
simple that a refactoring wouldn't buy us much.

We could do some additional DRY by creating a test monad that
abstracts away the differences between the two evaluation
implementations, and then rewriting most of the tests to use this
abstracted test framework. However, it's not clear whether it would be
worth the effort, as it would be a fairly significant amount of work
and would largely exchange boilerplate for boilerplate.

Note that this commit is intended to be a pure refactoring: no
functional changes have been made to the tests, nor have any tests
been added or removed.

Signed-off-by: Drew Hess <src@drewhess.com>