fix(primer-selda): switch to `novadiscovery/selda` fork
The `haskell.nix` updates in the previous commit broke the
`primer-selda`, due to `selda` not being compatible with GHC > 9.4. We
had been able to work around it with a patched upstream, but that
stopped working in the last commit. (Note that it had not been working
in the `nix develop` shell for some time.)

The helpful folks at Novadiscovery have forked `selda` to make it work
with GHC > 9.4. We adopt that fork here. Note that it also requires a
trivial change to how we create instances of `Selda.Relational` due to
this commit:

https://github.com/novadiscovery/selda/commit/1a40a9951c298f61dfaa46967f1991e3443ded14

Note that we still plan to replace `selda` (again), but this will do
until we have time for work.

Signed-off-by: Drew Hess <src@drewhess.com>