chore: Fix HLint
We had a lot of warnings about use of `maximum` etc. where HLint isn't smart enough to realise that we're using the safe versions for non-empty structures from `Data.Foldable1`. The simplest fix here is to move these in to `Foreword` so that we use them universally, and configure HLint to recognise this. While we're at it, to avoid export conflicts in `Foreword`, we do the same for `head` (which we never actually use).

Signed-off-by: George Thomas <georgefsthomas@gmail.com>