chore(primer-selda): build for Wasm
Note that the `primer-selda` tests build, but fail, because they try
to run `primer-sqitch` in a separate process, but `process` throws a
runtime exception under Wasm (on purpose):

https://github.com/haskell/process/blob/166971ffd5a750dd0b101f208f2ddc624d8d6997/System/Process.hs#L904

This is completely orthogonal to whichever Sqlite implementation we
choose, of course, and nothing to do with Selda. In fact, in theory,
Selda supports its own migration system, so we may be able to replace
`primer-sqitch` with Selda's own functionality. (That said, I believe
Selda's support for migrations is not particularly robust.)

In any case, if we want to use Sqlite in Wasm, this probably means the
end of the road for `sqitch`, and moving the database init/migration
into native Haskell code.

Signed-off-by: Drew Hess <src@drewhess.com>