style(primer-miso): Use node colour for its outgoing edges
The major change here is making edges actual children of their source node in the DOM, allowing the colour to be inherited. The separate "node-contents" is necessary for this, but adds more flexibility which could be useful in future anyway, e.g. if we want labels on nodes like we had in the JS version.

This commit also unintentionally solves in passing an issue where pattern edges were invisible due to bad z-indices.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>