style(primer-miso): Use custom fonts
These are taken from `primer-app`, our old JavaScript frontend.

We also copy licenses, and the section from the README about third-party code (including the part about vendoring, which isn't actually relevant to changes in this commit, but is appropriate due to the layout algorithm from diagrams).

Note that we don't yet use the "branding" font, but we might as well copy all the fonts over together.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>