style(primer-miso): Drop `outerPadding`
It's a bit pointless right now, and removing it allows us to simplify things. When we layout the UI properly with components in grids we'll likely want to do things differently anyway.

For now we apply some padding in to the whole canvas instead, via the stylesheet. Which at least means that hover rings don't overflow the screen.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>