refactor(primer-miso): Move some properties from stylesheet back inline
The plan was initially to go the other way, removing remaining `style_` tags, and working with computed values in the stylesheet somehow.

Instead, we note that uses of CSS properties mostly fall cleanly in to two categories, and use the stylesheet solely for the latter:
- Functionality: properties which are essential for the app to work.
- Aesthetics: properties which could in theory be overridden by a user-specific "theme". We can remove these entirely and have an app which is still functional, if ugly.

Another way to think about this is that if we were to separate out functionality in to React-style components, the former would be hardcoded to ensure reasonable encapsulation, and the latter configurable. This is a common pattern in libraries such as ReactFlow.

In some cases, it might make sense for something to be overridden even when it is necessary that it has _some_ value. For example, without setting `border-style: solid` on edges they can't be seen at all, but a theme could still override with `border-style: dotted !important`. See also the flexbox properties for `node-contents`. In these cases, we use inline styles to preserve the principle that the stylesheet-less app still works.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>