chore: remove primer-service
We're moving to a GHC-based web frontend implementation, so we don't
need primer-service anymore, and as we were the only user of it, we
can do this with impunity.

It would be reasonably easy to resurrect this API in the future, at
least the Haskell-friendly Servant bits, if we change our minds.
However, its unlikely that we would bring back the OpenAPI bindings,
as those required significant effort to maintain, and due to the
impoverished JSON-based serialization API, shed much of the type
safety and other benefits of developing applications in Haskell.
Having said that, we do still mention OpenAPI here and there, and even
keep most of our accommodations for it in `primer-api` (on purpose),
in case we or anyone else want to explore that path again in the
future.

Note that while I've tried to clean up the documentation to remove
references to the service, the HTTP API, etc., it's likely there are
still some references and other non sequiturs.

Signed-off-by: Drew Hess <src@drewhess.com>