fix: bump `ghc-wasm` flake to fix `wizer`
The version of `wizer` pulled in by this updated `ghc-wasm` flake
supports tail calls, which we now need.

(This `ghc-wasm` flake version breaks the Wasm build on macOS, or at
least it does on macOS 15.4.1, but we'll just have to deal with the
breakage for now, as the need for the updated `wizer` overrides a
working macOS build.

Signed-off-by: Drew Hess <src@drewhess.com>