feat (primer-miso): Add support for editing programs
This enables all actions on term definitions (but not typedefs), along with undo and redo.

Note that now that we need a full `App` in the model, it's harder to assert at the type level the invariant that we only work with programs which have been typechecked (`ExprT` rather than `Expr`, etc.). We might be able to refactor the core Primer library in order to create an `AppT`. This would likely cause a lot of churn, though it could help to illuminate when exactly we can rely on the typechecker having already been run. See the comment on `assumeDefHasTypeCheckInfo`, and note that we removed `tcBasicProg` and the "typecheck everything up front" comment on `Model`.

There is also a slight regression for now in that we lose the ability to view `Prelude` defitions, since the Primer API throws an error if we try to `MoveToDef` to a non-editable definition. This wasn't a problem before because we manipulated selections directly instead of using the proper API.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>