refactor(primer-miso): use `MisoString` where appropriate
The released version of Miso uses `MisoString = Text` on all platforms except JS, which is why we get away with using `Text` everywhere before this change. Miso's `HEAD` now uses `JSString` instead of `Text` for Wasm, so this future-proofs against that change, and should bring performance gains when the time comes.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>