refactor(primer-miso): Simplify selection actions
N.B. The way we use `setSelectionAction` is a bit silly really, because it deconstructs the selection, then applying the resulting `ProgAction` basically just reconstructs it and sets it as the program's selection. But this isn't really an issue in practice, and it's no worse than before this refactor anyway, only more obvious. Plus, `applyProgAction` does first perform some sanity checks for existence of the given definition etc.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>